Created a new Matrices class to replace the Matrices struct that was present in the CubeTransformDemo.
