+++ OpenGL/CubeTransformDemo.hpp
@@ -15,164 +15,7 @@
 #include <cmath>
 #include "ShaderLoader.hpp"
 #include "GLParams.hpp"
-
-enum Direction {
-    LEFT,
-    RIGHT,
-    UP,
-    DOWN
-};
-
-enum ScaleMag {
-    LARGER,
-    SMALLER
-};
-
-struct Matrices {
-private:
-    float rotate_x = 0.0f;
-    float rotate_y = 0.0f;
-    float rotate_z = 0.0f;
-    float move_x = 0.0f;
-    float move_y = 0.0f;
-    float scale_mag = 1.0f;
-    
-public:
-    Matrices() {
-        std::cout << "Construct: Matrices." << std::endl;
-    }
-    ~Matrices() {
-        std::cout << "Destruct: Matrices." << std::endl;
-    }
-    float rotation_x[16] = {
-        1.0f, 0.0f, 0.0f, 0.0f,
-        0.0f, cosf(rotate_x), sinf(rotate_x), 0.0f,
-        0.0f, -sinf(rotate_x), cosf(rotate_x), 0.0f,
-        0.0f, 0.0f, 0.0f, 1.0f
-    };
-    
-    float rotation_y[16] = {
-        cosf(rotate_y), 0.0f, -sinf(rotate_y), 0.0f,
-        0.0f, 1.0f, 0.0f, 0.0f,
-        sinf(rotate_y), 0.0f, cosf(rotate_y), 0.0f,
-        0.0f, 0.0f, 0.0f, 1.0f
-    };
-    
-    float rotation_z[16] = {
-        cosf(rotate_z), sinf(rotate_z), 0.0f, 0.0f,
-        -sinf(rotate_z), cosf(rotate_z), 0.0f, 0.0f,
-        0.0f, 0.0f, 1.0f, 0.0f,
-        0.0f, 0.0f, 0.0f, 1.0f
-    };
-    
-    float translation[16] = {
-        1.0f, 0.0f, 0.0f, 0.0f,
-        0.0f, 1.0f, 0.0f, 0.0f,
-        0.0f, 0.0f, 1.0f, 0.0f,
-        move_x, move_y, 0.0f, 1.0f
-    };
-    
-    float scaling[16] = {
-        scale_mag, 0.0f, 0.0f, 0.0f,
-        0.0f, scale_mag, 0.0f, 0.0f,
-        0.0f, 0.0f, scale_mag, 0.0f,
-        0.0f, 0.0f, 0.0f, 1.0f
-    };
-    
-    void scale(ScaleMag direction) {
-        switch(direction) {
-            case LARGER: {
-                if(scale_mag <= 0.8f) scale_mag += 0.05f;
-                break;
-            }
-            case SMALLER: {
-                if(scale_mag >= 0.1f) scale_mag -= 0.05f;
-                break;
-            }
-        }
-        scaling[0] = scaling[5] = scaling[10] = scale_mag;
-    }
-    
-    void translateX(Direction direction) {
-        switch(direction) {
-            case LEFT: {
-                move_x -= 0.05f;
-                break;
-            }
-            case RIGHT: {
-                move_x += 0.05f;
-                break;
-            }
-        }
-        translation[12] = move_x;
-    }
-    
-    void translateY(Direction direction) {
-        switch(direction) {
-            case UP: {
-                move_y += 0.05f;
-                break;
-            }
-            case DOWN: {
-                move_y -= 0.05f;
-                break;
-            }
-        }
-        translation[13] = move_y;
-    }
-    
-    void rotateX(Direction direction) {
-        switch(direction) {
-            case UP: {
-                rotate_x += 0.05f;
-                break;
-            }
-            case DOWN: {
-                rotate_x -= 0.05f;
-                break;
-            }
-        }
-        rotation_x[5] = cosf(rotate_x);
-        rotation_x[6] = cosf(rotate_x);
-        rotation_x[9] = -sinf(rotate_x);
-        rotation_x[10] = cosf(rotate_x);
-    }
-    
-    void rotateY(Direction direction) {
-        switch(direction) {
-            case LEFT: {
-                rotate_y += 0.05f;
-                break;
-            }
-            case RIGHT: {
-                rotate_y -= 0.05f;
-                break;
-            }
-        }
-        rotation_y[0] = cosf(rotate_y);
-        rotation_y[2] = -sinf(rotate_y);
-        rotation_y[8] = sinf(rotate_y);
-        rotation_y[10] = cosf(rotate_y);
-    }
-    
-    void rotateZ(Direction direction) {
-        switch(direction) {
-            case LEFT: {
-                rotate_z += 0.05f;
-                break;
-            }
-            case RIGHT: {
-                rotate_z -= 0.05f;
-                break;
-            }
-        }
-        
-        rotation_z[0] = cosf(rotate_z);
-        rotation_z[1] = sinf(rotate_z);
-        rotation_z[4] = -sinf(rotate_z);
-        rotation_z[5] = cosf(rotate_z);
-    }
-};
+#include "Matrices.hpp"
 
 class CubeTransformDemo {
 private:
