+++ OpenGL/Enumerations.h
@@ -1,0 +1,24 @@
+//
+//  Enumerations.h
+//  OpenGL
+//
+//  Created by Matt Finucane on 20/01/2017.
+//  Copyright © 2017 Matt Finucane. All rights reserved.
+//
+
+#ifndef Enumerations_h
+#define Enumerations_h
+
+enum Direction {
+    LEFT,
+    RIGHT,
+    UP,
+    DOWN
+};
+
+enum ScaleMag {
+    LARGER,
+    SMALLER
+};
+
+#endif /* Enumerations_h */
